works now with unittesting framework
